+++ src/layout/section/textFieldsSection/index.tsx
@@ -1,0 +1,2 @@
+export { TextFieldsSection, TextFieldsSectionProps } from './TextFieldsSection'
+export { TextFieldsSection as default } from './TextFieldsSection'