+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Unreleased
 
+- **[NEW]** Create `TextFieldsSection` to enable multiple `TextField` in the same row
 [...]
 
 # v41.2.0 (12/10/2020)